+++ lib/src/models.dart
@@ -11,3 +11,4 @@ export 'models/paginated_result.dart';
 export 'models/serializable.dart';
 export 'models/video_filter.dart';
 export 'models/channel_filter.dart';
+export 'models/search_filter.dart';