+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 - **Breaking**: Changed `HolodexClient::getVTuberCollabs` parameters to an optional `ChannelVideoFilter`
 - **Breaking**: Renamed `HolodexClient::getVideoMetadata` parameter `timestampComments` to `includeTimestampComments`
 - **Breaking**: Renamed `HolodexClient::getVideoMetadata` parameter `recommendationLanguages` to `languages`
+- **Breaking**: Changed `HolodexClient::searchVideos` parameters to an optional `SearchFilter`
+- **Breaking**: Changed `HolodexClient::searchComments` parameters to an optional `SearchFilter`
 - **Breaking**: Removed `VideoWithCommentsList`
 - **Breaking**: Removed `VideoWithComments`, use `VideoFull` instead
 - **Breaking**: Removed `VideoMetadata`, use `VideoFull` instead